Move telemetry enablement check to config
