+++ common/telemetry/grpc.go
@@ -8,7 +8,6 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
-	otelnoop "go.opentelemetry.io/otel/trace/noop"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/rpc/interceptor/logtags"
@@ -189,12 +188,4 @@ func (c *customServerStatsHandler) TagConn(ctx context.Context, info *stats.Conn
 
 func (c *customServerStatsHandler) HandleConn(ctx context.Context, stat stats.ConnStats) {
 	c.wrapped.HandleConn(ctx, stat)
-}
-
-func isEnabled(tp trace.TracerProvider) bool {
-	if tp == nil {
-		return false
-	}
-	_, isNoop := tp.(otelnoop.TracerProvider)
-	return !isNoop
 }
